- Update to new release: sendmail-8.12.3 - Patches removed that are now in the distribution - More MILTER dokumentation - sm-client.sh.sample have now 'restart' command. - Add standalone startupscript for all sendmail deamons - Update pkg-message
This version fixes a long-standing MIME (7 to 8-bit) conversion bug
and several smaller problems, e.g., a possible communication problem
between the MTA and libmilter, a bug in handling (invalid) addresses
containing 8-bit characters, a possible problem with small timeouts
being lost on slow machines if itimers are used, and the handling of
the 421 reply code and timeouts in the SMTP delivery code.  There are
two changes in the msp feature that may change the behavior: MX lookups
are turned on for the SMTP mailers and hence `[localhost]' is used as
default instead of `localhost', and confTIME_ZONE is set to USE_TZ in
submit.mc.  A complete list of changes can be found in the release notes.

+++ mail/sendmail/files/sm-client.sh
@@ -12,11 +12,20 @@ start)
 	echo -n ' sm-msp-queue'
 	;;
 stop)
-	kill `head -1 /var/spool/clientmqueue/sm-client.pid`
-	rm -f /var/spool/clientmqueue/sm-client.pid
+	if test -e /var/spool/clientmqueue/sm-client.pid
+	then
+		kill `head -1 /var/spool/clientmqueue/sm-client.pid`
+		rm -f /var/spool/clientmqueue/sm-client.pid
+	fi
+	;;
+restart)
+	if test -e /var/spool/clientmqueue/sm-client.pid
+	then
+		kill -1 `head -1 /var/spool/clientmqueue/sm-client.pid`
+	fi
 	;;
 *)
-	echo "Usage: ${0##*/}: { start | stop }" 2>&1
+	echo "Usage: ${0##*/}: { start | stop | restart }" 2>&1
 	exit 65
 	;;
 esac
